+++ nengo/node.py
@@ -1,4 +1,5 @@
 import warnings
+import inspect
 
 import numpy as np
 
@@ -7,8 +8,19 @@ from nengo.base import NengoObject, ObjView
 from nengo.exceptions import ValidationError
 from nengo.params import Default, IntParam, Parameter
 from nengo.processes import Process
-from nengo.utils.compat import is_array_like
+from nengo.utils.compat import is_array_like, getfullargspec
 from nengo.utils.stdlib import checked_call
+
+
+class OutputFnArgsValidationError(ValidationError):
+    def __init__(self, output, attr, node):
+        n_args = 2 if node.size_in > 0 else 1
+        msg = ("output function '%s' is expected to accept exactly "
+               "%d argument" % (output, n_args))
+        msg += (' (time, as a float)' if n_args == 1 else
+                's (time, as a float and data, as a NumPy array)')
+        super(OutputFnArgsValidationError, self).__init__(
+            msg, attr=attr, obj=node)
 
 
 class OutputParam(Parameter):
@@ -48,11 +60,7 @@ class OutputParam(Parameter):
             if node.size_out is None:
                 node.size_out = output.default_size_out
         elif callable(output):
-            # We trust user's size_out if set, because calling output
-            # may have unintended consequences (e.g., network communication)
-            if node.size_out is None:
-                result = self.coerce_callable(node, output)
-                node.size_out = 0 if result is None else result.size
+            self.coerce_callable(node, output)
         elif is_array_like(output):
             # Make into correctly shaped numpy array before validation
             output = npext.array(
@@ -70,23 +78,50 @@ class OutputParam(Parameter):
         return output
 
     def coerce_callable(self, node, output):
+        self.check_callable_args_list(node, output)
+        # We trust user's size_out if set, because calling output
+        # may have unintended consequences (e.g., network communication)
+        if node.size_out is None:
+            self.check_callable_output(node, output)
+
+    def check_callable_output(self, node, output):
         t, x = 0.0, np.zeros(node.size_in)
         args = (t, x) if node.size_in > 0 else (t,)
         result, invoked = checked_call(output, *args)
         if not invoked:
-            msg = ("output function '%s' is expected to accept exactly "
-                   "%d argument" % (output, len(args)))
-            msg += (' (time, as a float)' if len(args) == 1 else
-                    's (time, as a float and data, as a NumPy array)')
-            raise ValidationError(msg, attr=self.name, obj=node)
-
+            raise OutputFnArgsValidationError(output, self.name, node)
         if result is not None:
             result = np.asarray(result)
             if len(result.shape) > 1:
-                raise ValidationError("Node output must be a vector (got shape"
-                                      " %s)" % (result.shape,),
+                raise ValidationError("Node output must be a vector "
+                                      "(got shape %s)" % (result.shape,),
                                       attr=self.name, obj=node)
-        return result
+        node.size_out = 0 if result is None else result.size
+
+    def check_callable_args_list(self, node, output):
+        # not all callables provide an argspec, such as numpy
+        try:
+            func_argspec = getfullargspec(output)
+        except (TypeError, ValueError):
+            pass
+        else:
+            args_len = len(func_argspec.args)
+            if inspect.ismethod(output) or not inspect.isroutine(output):
+                # don't count self as an argument
+                args_len -= 1
+
+            defaults_len = 0
+            if func_argspec.defaults is not None:
+                defaults_len = len(func_argspec.defaults)
+
+            required_len = args_len - defaults_len
+            expected_len = 2 if node.size_in > 0 else 1
+
+            if func_argspec.varargs:
+                args_len = max(expected_len, args_len)
+
+            if not required_len <= expected_len <= args_len:
+                raise OutputFnArgsValidationError(output, self.name, node)
 
 
 class Node(NengoObject):